Move FollowingEditHelper to FeedDisplayHelper, as it is now used for List Management as well as Managed Followed Sites
