+++ client/reader/following-edit/list-item.jsx
@@ -12,7 +12,7 @@ const Icon = require( 'reader/list-item/icon' ),
 	Title = require( 'reader/list-item/title' ),
 	Description = require( 'reader/list-item/description' ),
 	Actions = require( 'reader/list-item/actions' ),
-	FollowingEditHelper = require( './helper' ),
+	FeedDisplayHelper = require( 'reader/lib/feed-display-helper' ),
 	decodeEntities = require( 'lib/formatting' ).decodeEntities,
 	FeedSubscriptionActions = require( 'lib/reader-feed-subscriptions/actions' ),
 	ReaderFollowButton = require( 'reader/follow-button' ),
@@ -91,15 +91,15 @@ var SubscriptionListItem = React.createClass( {
 			siteData = this.state.site,
 			feedData = this.state.feed,
 			iconUrl = siteData && siteData.get( 'icon' ),
-			displayUrl = FollowingEditHelper.formatUrlForDisplay( subscription.get( 'URL' ) ),
+			displayUrl = FeedDisplayHelper.formatUrlForDisplay( subscription.get( 'URL' ) ),
 			isFollowing = this.isFollowing(),
-			feedTitle = decodeEntities( FollowingEditHelper.getFeedTitle( siteData, feedData, displayUrl ) );
+			feedTitle = decodeEntities( FeedDisplayHelper.getFeedTitle( siteData, feedData, displayUrl ) );
 
 		const cardHeader = (
 			<div className="subscription-list-item__header-content">
 				<Icon>{ iconUrl ? <img src={ iconUrl.get( 'img' ) } alt="Feed icon" /> : null }</Icon>
 				<Title>
-					<a href={ FollowingEditHelper.getFeedStreamUrl( siteData, feedData, displayUrl ) }>{ feedTitle }</a>
+					<a href={ FeedDisplayHelper.getFeedStreamUrl( siteData, feedData, displayUrl ) }>{ feedTitle }</a>
 				</Title>
 				<Description><a href={ subscription.get( 'URL' ) }>{ displayUrl }</a></Description>
 				<Actions>
